Menu: extract the HUD tab into build_hud.cpp
Verbatim move of the HUD sections (text/indicator options, compass +
IMU source, color options, map overlay, mini-map/info-panel modules,
location, clock, menu position, legacy HUD). Built-in theme leaves come
from ctx.make_builtin_theme_leaves so System presets keep sharing them.
